chore(playwright): organize tests into directories
