+++ web/tests/e2e/admin/admin_oauth_redirect_uri.spec.ts
